Week 7 - Boarding Pass
Wrote and tested functions for getting and validating client's name, age, and departure date. Used the Amadeus API to retrieve real flight information which includes airport names, IATA codes, flight durations, prices, and segments.

Will need to review this code as a team to ensure it's clear and readable to all, and ddiscuss how we want to format and present this information to the user.